feature-mdr-39-terms-and-condition-page
