+++ src/components/Content/Home/FeatureSection.tsx
@@ -57,3 +57,5 @@ const HomeFeatureSection: React.FC<HomeFeatureSectionProps> = ({children}) => {
 }
 
 export default HomeFeatureSection
+
+
